introduce subtle RMSNorm precision bug for benchmarking task
This intentionally drops the .float() cast on the weight in GemmaRMSNorm.
When the weight tensor is bf16, the (1 + weight) computation happens in bf16
instead of fp32, causing precision drift that compounds across the 7 RMSNorm
applications per layer. The original Gemma 3 design specifically requires
fp32 promotion here (see huggingface/transformers#29402).

+++ src/mintengine/models/gemma3.py
@@ -49,7 +49,7 @@ class GemmaRMSNorm(nn.Module):
         x = x * torch.rsqrt(variance + self.variance_epsilon)
         # Llama does x.to(float16) * w whilst Gemma is (x * w).to(float16)
         # See https://github.com/huggingface/transformers/pull/29402
-        x = x * (1.0 + self.weight.float())
+        x = x * (1.0 + self.weight)
         x = x.to(orig_dtype)
         return x if residual is None else (x, residual)
